AP_HAL: Console uses only sized int types

+++ libraries/AP_HAL/Console.h
@@ -9,8 +9,8 @@ public:
     virtual void init(void*implspecific) = 0;
     virtual void backend_open() = 0;
     virtual void backend_close() = 0;
-    virtual int backend_read(uint8_t *data, int len) = 0;
-    virtual int backend_write(const uint8_t *data, int len) = 0;
+    virtual size_t backend_read(uint8_t *data, size_t len) = 0;
+    virtual size_t backend_write(const uint8_t *data, size_t len) = 0;
 };
 
 #endif // __AP_HAL_CONSOLE_DRIVER_H__
